refactor(analyzer): move validate_pl_variables_from_stmts to library
Now public at ogsql_parser::validate_pl_variables_from_stmts. Bridges
StatementInfo slices to per-PlBlock validate_pl_variables_* calls.
Part of #244.
